requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Audio processing and transcription
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0

# Text summarization and NLP
transformers>=4.30.0
sentencepiece>=0.1.99
nltk>=3.8.1
textblob>=0.17.1

# ML Tracking and versioning
mlflow>=2.8.0
dvc>=3.30.0

# Web interface
streamlit>=1.27.0

# Utilities
python-dotenv>=1.0.0
pydub>=0.25.1